+++ src/Symfony/Component/Messenger/Tests/Middleware/HandleMessageMiddlewareTest.php
@@ -11,13 +11,13 @@
 
 namespace Symfony\Component\Messenger\Tests\Middleware;
 
-use PHPUnit\Framework\TestCase;
 use Symfony\Component\Messenger\Envelope;
 use Symfony\Component\Messenger\Handler\Locator\HandlerLocator;
 use Symfony\Component\Messenger\Middleware\HandleMessageMiddleware;
+use Symfony\Component\Messenger\Middleware\StackMiddleware;
 use Symfony\Component\Messenger\Tests\Fixtures\DummyMessage;
 
-class HandleMessageMiddlewareTest extends TestCase
+class HandleMessageMiddlewareTest extends MiddlewareTestCase
 {
     public function testItCallsTheHandlerAndNextMiddleware()
     {
@@ -26,16 +26,13 @@ class HandleMessageMiddlewareTest extends TestCase
 
         $handler = $this->createPartialMock(\stdClass::class, array('__invoke'));
 
-        $next = $this->createPartialMock(\stdClass::class, array('__invoke'));
-
         $middleware = new HandleMessageMiddleware(new HandlerLocator(array(
             DummyMessage::class => $handler,
         )));
 
         $handler->expects($this->once())->method('__invoke')->with($message);
-        $next->expects($this->once())->method('__invoke')->with($envelope);
 
-        $middleware->handle($envelope, $next);
+        $middleware->handle($envelope, $this->getStackMock());
     }
 
     /**
@@ -46,13 +43,13 @@ class HandleMessageMiddlewareTest extends TestCase
     {
         $middleware = new HandleMessageMiddleware(new HandlerLocator(array()));
 
-        $middleware->handle(new Envelope(new DummyMessage('Hey')), function () {});
+        $middleware->handle(new Envelope(new DummyMessage('Hey')), new StackMiddleware());
     }
 
     public function testAllowNoHandlers()
     {
         $middleware = new HandleMessageMiddleware(new HandlerLocator(array()), true);
 
-        $this->assertNull($middleware->handle(new Envelope(new DummyMessage('Hey')), function () {}));
+        $this->assertNull($middleware->handle(new Envelope(new DummyMessage('Hey')), new StackMiddleware()));
     }
 }